Выделяет в компонент информацию об авторизации
