+++ project/src/pages/add-review/add-review.tsx
@@ -1,16 +1,16 @@
 import Logo from '../../components/logo/logo';
 import AddReviewForm from '../../components/add-review/add-review-form';
-import { Link, useParams, useNavigate } from 'react-router-dom';
+import { Link, useParams } from 'react-router-dom';
 import { useAppSelector } from '../../hooks';
 import NotFound from '../not-found/not-found';
 import { AppRoute } from '../../constants';
+import UserBlock from '../../components/user-block/user-block';
 
 export default function AddReview(): JSX.Element {
   const allFilms = useAppSelector((state) => state.films);
   const params = useParams();
   const id = params.id;
   const film = allFilms.find((movie) => String(movie.id) === id);
-  const navigate = useNavigate();
 
   if (!film) {
     return (
@@ -41,16 +41,8 @@ export default function AddReview(): JSX.Element {
             </ul>
           </nav>
 
-          <ul className="user-block">
-            <li className="user-block__item">
-              <div className="user-block__avatar" onClick={() => navigate(AppRoute.MyList)}>
-                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-              </div>
-            </li>
-            <li className="user-block__item">
-              <a href="/" className="user-block__link">Sign out</a>
-            </li>
-          </ul>
+          <UserBlock />
+
         </header>
 
         <div className="film-card__poster film-card__poster--small">
