+++ project/src/pages/my-list/my-list.tsx
@@ -1,6 +1,7 @@
 import Logo from '../../components/logo/logo';
 import FilmsList from '../../components/films-list/films-list';
 import { useAppSelector } from '../../hooks';
+import UserBlock from '../../components/user-block/user-block';
 
 export default function MyList(): JSX.Element {
   const allFilms = useAppSelector((state) => state.films);
@@ -11,17 +12,15 @@ export default function MyList(): JSX.Element {
       <header className="page-header user-page__head">
         <Logo light={false} />
 
-        <h1 className="page-title user-page__title">My list <span className="user-page__film-count">{favoriteFilms.length}</span></h1>
-        <ul className="user-block">
-          <li className="user-block__item">
-            <div className="user-block__avatar">
-              <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-            </div>
-          </li>
-          <li className="user-block__item">
-            <a href="/" className="user-block__link">Sign out</a>
-          </li>
-        </ul>
+        <h1 className="page-title user-page__title">
+          My list
+          <span className="user-page__film-count">
+            {favoriteFilms.length}
+          </span>
+        </h1>
+
+        <UserBlock />
+
       </header>
 
       <section className="catalog">
